Refactor: Rename MoesifMiddlewareExample -> MoesifNetCore2Example
Refactor: Rename MoesifMiddlewareExample -> MoesifNetCore2Example
