+++ MoesifNetCore2Example/MoesifNetCore2Example.csproj
